+++ e2e/testsuite/codec.go
@@ -25,6 +25,7 @@ import (
 	gmptypes "github.com/cosmos/ibc-go/v11/modules/apps/27-gmp/types"
 	icacontrollertypes "github.com/cosmos/ibc-go/v11/modules/apps/27-interchain-accounts/controller/types"
 	icahosttypes "github.com/cosmos/ibc-go/v11/modules/apps/27-interchain-accounts/host/types"
+	packetforwardtypes "github.com/cosmos/ibc-go/v11/modules/apps/packet-forward-middleware/types"
 	transfertypes "github.com/cosmos/ibc-go/v11/modules/apps/transfer/types"
 	v7migrations "github.com/cosmos/ibc-go/v11/modules/core/02-client/migrations/v7"
 	clienttypes "github.com/cosmos/ibc-go/v11/modules/core/02-client/types"
@@ -75,6 +76,7 @@ func codecAndEncodingConfig() (*codec.ProtoCodec, testutil.TestEncodingConfig) {
 	wasmtypes.RegisterInterfaces(cfg.InterfaceRegistry)
 	attestationstypes.RegisterInterfaces(cfg.InterfaceRegistry)
 	channeltypesv2.RegisterInterfaces(cfg.InterfaceRegistry)
+	packetforwardtypes.RegisterInterfaces(cfg.InterfaceRegistry)
 
 	// all other types
 	upgradetypes.RegisterInterfaces(cfg.InterfaceRegistry)